+++ requirements.txt
@@ -1,2 +1,3 @@
 yfinance
 pandas
+tabulate
